+++ INSTALL.md
@@ -54,14 +54,14 @@ After installation, register and use the data sources:
 
 ```python
 from pyspark.sql import SparkSession
-from python_data_sources.mcap import MCAPDataSource
+from src.python_data_sources.mcap import MCAPDataSource
 
 spark = SparkSession.builder.getOrCreate()
 spark.dataSource.register(MCAPDataSource)
 
-df = spark.read.format("mcap") \
-    .option("path", "/path/to/data.mcap") \
-    .option("numPartitions", "4") \
+df = spark.read.format("mcap")
+    .option("path", "/path/to/data.mcap")
+    .option("numPartitions", "4")
     .load()
 ```
