+++ Makefile
@@ -2,11 +2,13 @@ all: clean dev fmt
 
 clean:
 	rm -fr .venv clean .pytest_cache .ruff_cache .coverage coverage.xml
+	rm -fr **/*.pyc
 
 .venv/bin/python:
 	pip install hatch
 	hatch env create
-	hatch run pip install "python-data-sources[all]"
+	hatch run pip install ".[all]"
+	touch .venv/bin/python
 
 dev: .venv/bin/python
 	@hatch run which python
