+++ pyproject.toml
@@ -60,27 +60,21 @@ Repository = "https://github.com/databricks-industry-solutions/python-data-sourc
 Issues = "https://github.com/databricks-industry-solutions/python-data-sources/issues"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src"]
-
-[tool.hatch.build.targets.wheel.sources]
-"src" = "python_data_sources"
-
-[tool.hatch.build.targets.sdist]
-include = [
-    "/src",
-    "/tests",
-    "/README.md",
-    "/LICENSE.md",
-]
+packages = ["src/python_data_sources"]
 
 # =============================================================================
 # Hatch Environments - One test environment per submodule
 # =============================================================================
 
 [tool.hatch.envs.default]
+python = "3.11"
+
+path = ".venv"
+
 dependencies = [
     "black~=24.8.0",
     "databricks-labs-pylint~=0.5",
+    "databricks-sdk",
     "mcap>=1.3.0",
     "mcap-protobuf-support>=0.5.3",
     "paho-mqtt>=2.1.0",
@@ -89,6 +83,8 @@ dependencies = [
     "pylibjpeg[all]>=2.0.1",
     "pylint~=3.3.1",
     "pylint-pytest==2.0.0a0",
+    "pyspark>=4.0.0.dev1",
+    "pyarrow>=14.0.0",
     "pytest>=8.0.0",
     "pytest-cov>=4.0.0",
     "pytest-mock>=3.10.0",
