--- src/mcap/__init__.py
@@ -1,13 +1,0 @@
-"""MCAP data source for PySpark."""
-
-from python_data_sources.mcap.mcap_datasource import (
-    MCAPDataSource,
-    MCAPDataSourceReader,
-    RangePartition,
-)
-
-__all__ = [
-    "MCAPDataSource",
-    "MCAPDataSourceReader",
-    "RangePartition",
-]
