+++ src/python_data_sources/mcap/__init__.py
@@ -1,0 +1,7 @@
+"""MCAP data source for PySpark."""
+
+__all__ = [
+    "MCAPDataSource",
+    "MCAPDataSourceReader",
+    "RangePartition",
+]
