+++ tests/conftest.py
@@ -22,7 +22,11 @@ def spark() -> SparkSession:
     Create a SparkSession for testing.
     The session is reused across all tests in the session.
     """
-    spark_session = SparkSession.builder.master("local[2]").appName("python-data-sources-tests").getOrCreate()
+    spark_session = (
+        SparkSession.builder.master("local[2]")  # pylint: disable=no-member
+        .appName("python-data-sources-tests")
+        .getOrCreate()
+    )
     spark_session.sparkContext.setLogLevel("WARN")
     yield spark_session
     spark_session.stop()
