+++ tests/mcap/conftest.py
@@ -6,7 +6,7 @@ import pytest
 from pathlib import Path
 from pyspark.sql import SparkSession
 
-from python_data_sources.mcap import MCAPDataSource
+from src.python_data_sources.mcap import MCAPDataSource
 
 
 @pytest.fixture(scope="session")
